% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_coef_via_mm.R
\name{comp.B.MM.c}
\alias{comp.B.MM.c}
\title{Title comp.B.MM.c
compute the B-spline coef matrix B using MM algorithm via C code}
\usage{
comp.B.MM.c(
  ztau,
  Y,
  tau,
  breaks,
  basis.order,
  n,
  p,
  maxiter = 1000,
  tol = 10^-8,
  epsilon = 0.001
)
}
\arguments{
\item{ztau}{transformed preditor variable}

\item{Y}{response variable}

\item{tau}{quantile levels}

\item{breaks}{breaks for B-splines, including starting and finishing points of the interval}

\item{basis.order}{order of B-splines}

\item{n}{number of observations}

\item{p}{dimension of covariate}

\item{maxiter}{maximum number of iteration during MM algorithm}

\item{tol}{tolerance for convergence}

\item{epsilon}{parameter used in MM, the default value is 0.001}
}
\value{

}
\description{
Title comp.B.MM.c
compute the B-spline coef matrix B using MM algorithm via C code
}
