% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_defined_functions.R
\name{gen_bspline}
\alias{gen_bspline}
\title{Title gen_bspline
generate b-splines based on tau}
\usage{
gen_bspline(tau, breaks, degree)
}
\arguments{
\item{tau}{quantile level seq}

\item{breaks}{breaks for B-splines}

\item{degree}{basis order of B-splines}
}
\value{
vector
}
\description{
Title gen_bspline
generate b-splines based on tau
}
