% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_example.R
\name{gen_x}
\alias{gen_x}
\title{Title gen_x
example for generate preditor variable (drawn from exponential distribution exp(1))}
\usage{
gen_x(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{dimensions of covariate}
}
\value{
matrix
}
\description{
Title gen_x
example for generate preditor variable (drawn from exponential distribution exp(1))
}
