% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_example.R
\name{gen_y}
\alias{gen_y}
\title{Title gen_y
example for generate predictor y (a nonlinear or linear form)}
\usage{
gen_y(x, beta0, gamma0, e, v)
}
\arguments{
\item{x}{preditor variable}

\item{beta0}{initial value of coefficient beta}

\item{gamma0}{initial value of coefficient gamma}

\item{e}{epsilon(error term)}

\item{v}{index for nonlinear part}
}
\value{
vector
}
\description{
Title gen_y
example for generate predictor y (a nonlinear or linear form)
}
