% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_defined_functions.R
\name{phi}
\alias{phi}
\title{Title phi}
\usage{
phi(y, z_tau, tau, b, n)
}
\arguments{
\item{y}{response variable}

\item{z_tau}{transformed preditor variable}

\item{tau}{quantile levels}

\item{b}{coefficient value}

\item{n}{number of observations}
}
\value{
number
}
\description{
Title phi
}
