% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_lof_statistic.R
\name{Un.b.c}
\alias{Un.b.c}
\title{Title Un.b.c
calculate the bootstrap lack-of-fit test statistic}
\usage{
Un.b.c(x, x.b, phi, phi.b)
}
\arguments{
\item{x}{original sample}

\item{x.b}{boostrap sample of x}

\item{phi}{results of funciton Phi with respect to x, in vector form}

\item{phi.b}{results of funciton Phi with respect to x.b, in vector form}
}
\value{
number
}
\description{
Title Un.b.c
calculate the bootstrap lack-of-fit test statistic
}
