% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_lof_statistic.R
\name{Un}
\alias{Un}
\title{Title Un
calculate the lack-of-fit test statistic}
\usage{
Un(x, phi)
}
\arguments{
\item{x}{predictor variable}

\item{phi}{results of funciton Phi with respect to x, in vector form}
}
\value{
number
}
\description{
Title Un
calculate the lack-of-fit test statistic
}
